finish off support for using loan term details of a loan to determine interest due rather than inferring it from the number of repayments and repaid every information. This allows better support for loan schedules where the first repayment is irregular (e.g. not same timescale as repaid every)

+++ mifosng-provider/src/main/java/org/mifosng/platform/loanschedule/domain/PmtCalculator.java
@@ -51,7 +51,7 @@ public class PmtCalculator {
 		double interestRateFraction = periodInterestRateForLoanTermPeriod.doubleValue();
 		
 		double futureValue = 0;
-		double numberOfPeriods = loanTermFrequency;
+		double numberOfPeriods = loanTermFrequency.doubleValue();
 		double principalAmount = principal.getAmount().multiply(BigDecimal.valueOf(-1)).doubleValue();
 		
 		// work out the total payment (principal + interest components) for each installment due.
